12/26/2018 5PM all-users-admins-functions-added Samy

+++ public/admin/super/index.php
@@ -1,3 +1,5 @@
+
+
 <?php
 $page = 'dashboard';
 require_once('../../../private/initialize.php');
@@ -6,46 +8,48 @@ deny_user_access();
 deny_client_access();
 deny_admin_access();
 ?>
-<div class="container">
-  <a href="new.php">New Admin</a>
-  <table class="table">
-    <thead>
-      <tr>
-        <th scope="col">#</th>
-        <th scope="col">Name</th>
-        <th scope="col">Username</th>
-        <th scope="col">Email</th>
-        <th scope="col">Operations</th>
-      </tr>
-      <?php
-      //Get all categories from database
-      $admins = Admin::find_all();
-      foreach ($admins as $admin) {
-        echo "<tr>";
-        echo "<td>".$admin->getId()."</td>";
-        echo "<td>".$admin->getFullname()."</td>";
-        echo "<td>".$admin->getUsername()."</td>";
-        echo "<td>".$admin->getEmail()."</td>";
-        echo "<td>"
-        ."<a href='edit.php?id={$admin->getId()}'>". "  Edit " ."</a>"
-        ."<a href='delete.php?id={$admin->getId()}' "
-        ."onclick='return confirm(\"Are you sure?\")' >". " -  Delete" ."</a>"
-        ."</td>";
-        echo "</tr>";
-        //print_r($cat);
-      }
+<!--================Services Area =================-->
+<section class="services_area p_120">
+  <div class="container">
+    <div class="main_title">
+      <h2>Staff</h2>
+    </div>
+    <a href="new.php">New Admin</a>
+    <table class="table">
+      <thead>
+        <tr>
+          <th scope="col">Name</th>
+          <th scope="col">Username</th>
+          <th scope="col">Created</th>
+          <th scope="col">Updated</th>
+          <th scope="col">Email</th>
+          <th scope="col">Operations</th>
+        </tr>
+        <?php
+        //Get all categories from database
+        $admins = User::find_all_admins();
+        if(!empty($admins)) {
+          foreach ($admins as $admin) {
+            echo "<tr>";
+            echo "<td>".$admin->getFullname()."</td>";
+            echo "<td>".$admin->getUsername()."</td>";
+            echo "<td>".$admin->getCreatedAt()."</td>";
+            echo "<td>".$admin->getUpdatedAt()."</td>";
+            echo "<td>".$admin->getEmail()."</td>";
+            echo "<td>"
+            ."<a href='delete.php?id={$admin->getId()}' "
+            ."onclick='return confirm(\"Are you sure?\")' >". " -  Delete -" ."</a>"
+            ."</td>";
+            echo "</tr>";
+          }
+        } else {
+          echo "No admins to be shown";
+        }
+        ?>
 
-
-      ?>
-
-    </thead>
-
-
+      </thead>
+    </table>
   </div>
-  <script type="text/javascript">
-      $('.confirmation').on('click', function () {
-          return confirm('Are you sure?');
-      });
-  </script>
-</body>
-</html>
+
+</section>
+<!--================End Services Area =================-->
